+++ NEWS.md
@@ -4,6 +4,7 @@
 Install the development version with: `install_github("rstudio/reticulate")`
 
 - Improved filtering of non-numeric characters in Python / NumPy versions 
+- Added `py_func()` to wrap an R function in a Python function with the same signature
 
 
 ## reticulate 1.6 (CRAN)